refactor controller as per James request
